"3:05pm. Done with the bath. What I've done just now is gotten Anatomy For Sculptors.
I think I am just losing my nerve here. Yeah, there is nothing wrong with getting a job. There is nothing wrong with not wanting to spend all my time drawing or sculpting the same thing over and over again. It is not wrong to be nervous about my prospects as a writer or being desirous of higher income and safety that programming would provide me.

But there is nothing wrong with wanting to learn art either. There is always time to learn something. I might not have time to do an untold number of models, but I do have time to do a few of them.

Like I said, I can spend a month doing this.

4:05pm. What is the problem then?

Anime.

I am being overwhelmed by it.

Stylized art seems easier at first glance, and for bad artists it in fact is. It is true that it is a shortcut in all sorts of ways, but what it isn't a shortcut is for learning.

It is always a mistake to try to grasp the feeling directly.

What I should be doing is a starting with realistic bodies and simplifying them down to anime.

CSP models aren't good for this. Just what will I do with those dull hollow eyed models.

Once I understand exactly what my charas should be in 3d, that is when I am ready to actually draw them. I'll do the anime face in 3d and come to understanding of what the various expressions should be in 3d for any kind of perspective.

4:10pm. It is not bad to obsess about realism at this stage. I can easily smooth the model and remove detail after I am done with that. That is the point of stylization.

This is the direction I picked for my env art. It is time to replicate the same thing my chara art.

In a way what I've done with Scatter 5 with the city scene is what I should have done back in January instead of going to Houdini. An expert can do in a day what would take a novice months. If I had to do the flower scene today, I could easily do it in Blender in just a few days in a far simpler fastion than in Houdini.

4:15pm. Forget about the shitty technique of that VR tuber guy. Trying to do it like him is absolutely the wrong way to go. I'd honestly respect him more if he started with an already finished model and lightly adjusted it. That would have been faster as well.

4:25pm. If I focus on learning proper anatomy first, that will give me an objective measure for how good or bad my model is. What do I know how good a random anime face is?

4:30pm. Methodically going through that book will also give me a reason to redo the base mesh. Last time I was just immitating Flycat, but going by the book is the proper process.

Forget anime, let me prend my goal is to make a bunch of realistic characters.

Come to think of it, there should be some face2anime GANs. It might be worth looking into them once I have the proper model.

4:35pm. One thing I want to do is to make a skeleton as well. This is a good opportunity to do so.

5:15pm. I've decided.

I do not want to pursue the Heaven's Key plan while seriously expecting it will give me money. The value of Heaven's Key is that it cured me of my obsession with poker and made me consider more broadly what could be possible with AI.

Post-Singularity I'll be making stories such as it all the time.

As a human, what is really important is AI itself.

But at the same time it pissed me off not being able to do an anime character. 2022 is my vacation. If I want to spend a few months just sculpting the human body then that is no problem.

As far as environments are concerned, I am feel quite proficient at them now.

Considering how much effort I put into this in 2022, it should be no wonder. But chara modeling has only a single month invested into it.

5:25pm. I can't be mogged by some dumb Youtuber performance art.

If I want money I should just go and get it, and if I want to do art, I should just do art.

5:55pm. 30/228. It's an interesting book. Despite it being mostly pictures, I am making slow progress through it.

I feel like closing here.

I really should reconsider my path. There really is a lot of difference between being able to make a decent picture in under a minute like an AI would taking several days minimum like I do.

I said I would rather make 1k as an artist vs 10k as a programmer, but I should reconsider how sane that really is. Especially considering it is questionable if I am going to get to 1k in the first place. Being humple about my desires does not mean I will fulfil them, so why not have big desires to begin with.

I was wrong to want to rock the online gambling dens, I'd be better off trying to crash the anime industry instead.

6pm. Being good at art is not complicated. Can you do a house, can you do a tree, can you do a car...

You go through the list until you can do everything efficiently. If you can do all that you imagine, then you are good at it.

I'll take aim at the human body again, but I won't bother with struggling against the pros or any of that nonsense. I know I have a 5/5 skill in which I am greatly talented. So there is no point in agitating over being good over something I have low aptitude for. I should just internalize the body, and then move on to the next thing whether that be programming, music or Heaven's Key."