"7:50am. I came back from doing the chores 10m ago and have just been chilling. I am not sure when I got up, but it must have been something like 6am.
8am. Ugh, let me go to bed. I do not feel like even reading manga right now.

9:45am. I was in bed less than I thought. But I do feel less groggy now.

Regarding the sculpting task, I am starting to feel a bit inspired. I'll rant about it later. Let me chill for now.

10:55am. Let me have breakfast here. I feel hung over right now despite haven't having a drop of alchohol.

11:40am. Done with breakfast. Let me finish ch 636 of Legendary Mechanic and I will get started. I want to do some icebreaking.

12:05pm. Done with grapes as well. I've revived a little.

...Ok, let me rant a bit.

When I went to bed at 8am, my thoughts wondered for a bit and then I started thinking about sculpting and got some motivation.

It weird when I think about it. I was really proud for getting the base mesh looking as well as it did last year, but now that I am analyzing my overall worflow, I see that it is all over the place.

This time, my project rather than being a sulpt of a body is a skeleton.

With that goal in mind, I've comparing my old sculpting techniques from last year with my recent ones and I can see immediatelly what the problem is.

I visualize the skeleton in my mind, and immediatelly get lethargic, but once I replaced all the bones with boxes, I start to feel a sense of energy.

This is what my problem was last time - in 2021 I was just starting out and everything was overwhelming. I was just moving along while feeling things out.

12:25pm. I feel so tired again, but let push on with the rant. Maybe after that I'll take a nap for the rest of the day.

What I want to say is - last year I struggled a lot just to grasp the basic things. But right now I am thinking how I could make the skeleton out of tiny boxes and am thinking: 'This won't be hard.'

After that, it is not a big deal to union the them and go into sculpting to make the polished result.

I used to feel that the hard part was in the tiny details, but now I think that the initial block out is where the 80% of the work is.

I want to test this theory. Last year I fiddled so much, a full month just to get the base mesh done.

But now what I want to do is bring my organic sculpting inline with my hard surface skills. It needs to be simple and straightforward.

I have a vision of what the ideal workflow should be. You start out rough with big boxes, and gradually subdivide them and add detail.

Imagine if I tried doing the skeleton like I had the body last year. I'd never get anywhere!

Imagine snake hooking the ribs off the sternum! That would be such a mess. And yet that is how I did the fingers, and yes it was hard.

12:30pm. There are better and worse ways of doing things. By far the worst way is the one used by the anime head guy.

The best way is what I am suggesting here - assemble the body out of primitives, boolean them together and then use sculpting to integrate the parts and do the detailing.

I know Flycat did this on one mesh, but he showcased various different ways of doing it, some of them worse.

Yan does it like I've suggested and does not mess around.

12:35pm. At my present level, I have zero reason to be proud of sculpting. This is in contrast to my Moi work.

If I do the base mesh again and take a month to do it, that would be a disaster.

Instead I will take my hard surface skills, and make them an integral part of the block out phase, making it so that by the time I get to sculpting only the easy stuff is left.

12:40pm. Also doing it like this will mark a change in attitude. I will do this while studying the Anatomy For Sculptors book.

I am going to internalize and master each and every part of the body as I go along.

12:40pm. Come to think it the way he did the base mesh is by starting from a single cube, extruding and subdividing it. When I did my own imitation I took the same approach.

Back then I contrasted that approach with doing the individual block out pieces and booleaning them. The former approach felt more succinct to me, but now I see that I was wrong.

The proper way to do it is to piece it together, only that approach gives you the flexibility of doing sketching in 3d.

The single cube deformation approach gets you into the mindset of thinking about the body as a single large mass and that is not good.

That way of doing it, while not extreme as the anime head guy's is a performance. It is no different from Kim Jung Gi xeroxing stuff from his mind directly onto paper.

You absolutely can't immitate such a thing if you are a regular person. Practice is not enough.

I'd rather find the right workflows that anybody could immitate. Better tools, better techniques - that is how you get better as a programmer, and that is how I intend to get better in art as well. Practice matters too, but without the tools and the techniques to cultivate, you'd only reach a low level.

Well, something like drawing is fundamentally about putting lines on paper, so with strong talent and experience you can reach astonishing heights of skill like Kim.

But in programming this would not work. Imagine if you were the best assembly programmer - you'd still be far behind in productivity compared to a mediocre Python programmer.

12:50pm. Some tools are easier and harder to use.

The manual mode of the Scatter 5 plugin is revolutionary for that reason - even if I don't use it for a few years and then come back, it would be very easy to remember what the scatter brushes do. It is like trying to ride a bike again. But getting a grasp on geo nodes and Houdini's scattering would be a pain in the ass.

I was simply ignorant.

I had no idea that in ray tracing mode Blender could display more poly than using the workbench engine. What a surprise that was. Now that I know the tricks to scattering, I have much less need for Clarisse or Houdini.

1:25pm. Had to take a short break.

Let me conclude the rant - it annoys me to have to work so hard.

Who feels like doing dozens of illustrations, compose music and do writing for a negligible chance to reaching financial security.

I do not want to be burdened with all of that.

And I do not want to burden myself with sculpting a bunch of charas for the next few months.

Who cares about all of that.

In programming I managed to create Spiral by taking it one step at a time.

In art, instead of going forward and trying to churn out things faster, what I want to do is take a step back and cover the missing fundamentals.

In past 9 months I've done a great big tour, so it will be almost a letdown what I am going to do, but it is essential if I want to bring my sculpting to the next level.

Rather than grand ambitions, what I need to cultivate here is a dissatisfaction with a single aspect of my art and laser focus on it.

Regarding the anime style, I've had an idea. Figures.

https://store.crunchyroll.com/collections/figures

It slipped my mind, but I bet I could find some good 3d references for them somewhere.

This is what I've been missing as my drawing target - a complete 3d model. The CSP model is not enough and is just leading me down the wrong track. If I want to learn to draw 2d anime, what I should use as references are figures such as these.

1:45pm. Let me go to bed. I'll step away from the screen for the rest of the day, till about 5-6pm and then watch Birdie Wing and play Genshin Impact after that. Fuck this shit, I can always get a job if I run out of time for real. Until then I should just take it easy and focus on one thing at a time. So what if I can't get to a full pro level so easily, I've already come a decent amount. I am certainly better than when I started.

Maybe tomorrow I'll manage to cultivate enough desire to begin for real."