"8:30am. I got up at 6:30am and have been doing chores since then. I need some time to cool down.
I've been thinking about the job quest. Whether it is worth it for me to find a job all depends on the cost of training a system like DALLE.

https://news.ycombinator.com/item?id=31228710
> One thing I’ve not seen mentioned/asked about DALL-E 2, which I’m really, really intrigued about is what hardware its typically running on and the average time it takes for image generation? Maybe I’ve somehow missed it, but I’m really intrigued by what their hardware set up is and how long it on average takes to generate the eight images from a users text input? I’m guessing it’s really quick, but I honestly have no idea… How does this system scale? Typical cost per image generated?

8:50am. https://twitter.com/alexjc/status/1348267025320304641
> My current estimate for training DALL·E is around $300k-$500k, since it's around 15x smaller than GPT-3 and that's estimated to have cost $4.6M. DALL·E has an additional VQVAE2 decoder, which may increase training times...

Forget this. I am not going to work for 5 years just to gamble on a single training run.

8:55am. Yeah, the Singularity in this timeline is so far is incredibly bootleg. Let me post a link.

https://semianalysis.substack.com/p/nvidia-in-the-hot-seat
> Intel Habana, Graphcore, Google TPU, and Nvidia A100 Compared In AI Training

I saw this 2 days ago, but I missed my chance to post it so let me do it now.

I really expected the AI chips to be able to beat the GPU 10x initially and 100x at the end of the road, and this is without memristor breakthroughs.

Instead I have this trash where they outright fail to beat the GPUs on most tasks, and when the finally win, it is something trivial like 30%. I do not buy the excuse the software is the problem, NNs aren't that hard to program.

Just what can I do with this? I'd like to get a job so I can actually buy something with that money. But this certainly is not worth buying. So I do not really need a job after all.

I wish I could do something about my mother's cancer situtation, but cancer is not something I can deal with using the power of money either. Even if I cryofroze her once the situation got bad enough, I'd need post Singularity tech to resurrect her. And is we are talking post-Singularity, things like asking whoever is running the universe for backups, and even something like resimulating the universe becomes viable. I am not going to bother thinking too seriously about these matters.

So in the end, I could only get a better GPU and CPUs with the money from a job I would get. Not useless, but it is not a big deal either given the Moore's Law breakdown.

Getting a job would only matter if it would get me the kind of hardware I'd need to train DALLE cheaply.

9:05am. Even if I got a job, I can't bet on the hardawre improvements bringing the figure down significantly in 5 years time.

There are basically only two things that matter at this point:

* Revolutionary algorithmic improvements that could make RL viable. If I could make guaranteed money via RL, I could then reinvest it and form a positive feedback loop.
* Memristor breakthroughs leading to huge hardware improvements. This might be needed to enable the algo revolution. Also would actually allow me to cultivate DALLE style systems cheaply even without revolutionary algo advances.

What are the chances of the two points happening in the next 5 years? Who knows.

Even if the second point happened today, it would still take years for the devices to be commercially viable. Realistically, it could happen in the next few years, but it is the kind of thing that I need to see to believe it first. I won't forget how I got suckered the last decade by the memristor hype.

9:10am. The first point would have the most immediate effect and would make my poker plan viable even with today's hardware. Since it is just knowledge, once it is shared it would have tremendous impact on RL.

It is a big mystery to me if the brain algorithm could be run efficiently on GPUs, or if it requires novel hardware. On the issue of whether intelligence could be optimally run on GPUs or whether it requires neuromorphic chips, either side seems equaly plausible to me.

9:15am. Since the situation is like this, I am helpless regarding anything I want to do. I can only sit tight and work on my hobbies.

Just forget this thing. The pursuit of Singularity might only become viable 10 years from now, and 10 years from now who knows what will happen. I could be dead for all I know. I am just human. 10 years is perfectly long enough to develop and die cancer myself or get hit by a truck while taking out the trash.

9:20am. 500k to train DALLE is not exorbiantly expensive. It is not 50M. But no starving artist such as myself could afford such a figure anyway.

9:25am. Forget this. I have plenty of time to practice drawing, music, 3d, write and finish the C backend for Spiral.

This NEET existence is something I've earned by finishing school a decade and a half ago, so I might as well stay snug.

Let me read Baki here. Egg of the Elf is out as well. I'll try to put in a few hours into doing 3d today, just placing boxes in the right spots. Right now I am still groggy so I do not feel like it. Maybe I'll take a nap."