"https://www.nextbigfuture.com/2020/11/memristors-and-hpe-machine-focused-computer-research-seems-almost-dead.html
Sad. Forget about memristors for now.

11:15am. Done with breakfast. Let me try working on that skeleton for a bit. I want to go to bed, but am too agitated to do that. I wouldn't be able to get any good rest at this point.

11:20am. I created the Eaten in Darkness - Gray Goo folder and saved the default cube as `skeleton.blend`.

Now what is next. I've resized the cube a bit, but I am not sure what the width and the depth of a human body is.

Let me open the anatomy for sculptors book.

11:30am. Page 9 of 220 has a pretty good skelly profile image from both sides.

11:40am. Interesting that you can change between objects while in edit mode using Alt Q. I don't think that a year ago Blender had this feature. It will be convenient.

At any rate, I am starting to feel it a bit. Just one big box for the body does not do much for me, but now that I am trying to figure out the proportions of the head I am getting into it. Usually what I'd do is important the ref image directly into Blender and set it on the background, but I want to move away from that.

I want to try doing the skeleton using my own aesthetic sense while going back and forth between the book and Blender. That should lead to better learning.

11:45am. Focus on Blender, not on this journal. Get back to work me.

12:15pm. Oh yes, feet. Right now I am just trying to stay awake.

But besides that I am making steady if very slow progress.

12:50pm. Let me take a short break here and then I will write down my thoughts.

1:25pm. I back. Anyway, I am done with the first block out phase. This bring back memories to the base mesh I did. I found it quite hard to immitate what Flycat was doing back then. For example when he did the back and the neck, getting the exact angle was difficult. But now I do not think that matters much.

What this version is, is merely a launch off point for the subsequent phase. I do not need to concern myself too much with details and angles.

If I could see the compled image superimposed on these primitives then yes, I'd be able to see where I am going wrong. But even so this is good enough to get a sense of the proportions. With just a cube I had a very rough view, but with this version, I can see much better. I can see the head, the arms, the legs, the chest, the buttocks and the abdomen. They are roughly in the right place.

Now, if it was Yan or Flycat, they'd jump into the sculpting phase.

But I have a better idea - why not go for another block out phase? There is no reason why it has to be only a single one. Now that I have a body made of large primitives, why not go for a body made of larger primitives and use the previus phase as a scaffolding?

This is the way to do it. It gets to the essence of sketching in 2d art.

If I follow this path, I think for models that I am familiar with I should be able to make them from imagination. This is like the holy grail of 2d art, but in 3d it would not be a big deal.

But you do need to follow the proper workflow. Imagine trying to do it like the anime head guy - unless you can xerox it like Kim Jung Gi you'd never get anywhere using such a method.

1:35pm. The block out I have right now is suitable for a body, but for the next step, I'll want to move to modeling individual bones. I won't bother detailing them, and some composites like the ribcage and the spine, I won't bothe individuating, but otherwise, the final product of this version should be recognizable as a low poly skelly.

1:40pm. I do not feel like doing it now. Today was just an icebraker, and I keep yawning from lack of sleep.

Yesterday I went to bed too late given how early I am getting up. At slightly before 11pm. This beats my usual 1-2am, but I guess I am still not getting enough sleep.

There is no helping it, doing these chores so early in the morning for the sake of winter preparation will distrupt my rhythm. Maybe I'll get used to it by tomorrow.

Let me go to bed for a while.

Even though I won't be doing the work that I should, I will stick to my principles and not play games or read novels or manga. This will help with building my determination.

A pro would breeze through this, but I can only treat this project like I would a programming project. In programming I have a hard time starting on new parts too, and I always just take the time to brainstorm it. This helps me get motivated to do the work as well."