"8:45am. I am back. I've been stacking wood for 3h now. Let me chill for a while, maybe I'll even take a nap. Yesterday was strange, I actually fell asleep when I went to bed. This never happens to me in the middle of the day, that is how tired I was. In this state, my desire to do art is quite low.
Initially I thought it was like programming, but in programming you can think about things ahead of time and just come to expend that creative urge. But art is quite laborious. Well, I'll still do it. I am going to finish this project and do some chara modeling for the next month or so. After that who knows. Maybe I will take the easy route of taking a job. It would certainly be less work than Heaven's Key.

8:50am. Yeah, let me go to bed. Screw this.

9:35am. No, I can't sleep unfortunately. I am not like my dad who can lie down and go to dreamland right away. Yesterday it only worked because I was so tired after the art session. I guess I not quite tired enough.

I am thinking of doing the chores from 6-9pm as well. If I do it 5-6h a day, I could be done very quickly, in like 3 days.

I have a bout 4.5 rows left. In the last two days, I did 1.3 rows roughly, and today I did 1 whole row.

Ok, I'll try doing it like that. Let me just chill a bit here and then I will do some art. The sooner I get this over with, the sooner I can get back to having proper sleep.

10:10am. Let me set Bastard to download. While that is going on, let me see if I can make progress on that model.

10:15am. Mhhh, the image is a bit low res. I can't tell properly how many bones the toes have. At rate, let me start from there.

10:20am. Let me go to the toes chapter.

10:30am. Ok...

It seems hugely complicated, but all I have to do is make sure it makes sense in two perspectives and I am 90% there.

3d is not as hard as it seems. Now me, put those boxes there.

The book has really good pages on toes.

10:50am. This actually is hard.

10:55am. Sure, being tired is inhibiting my learning, but this is by no means easy.

Right now I am really glad that I've gone with the plan of doing gradual block out. If I tried modeling the individual bones directly, I'd definitely gotten the proportions wrong and failed at the task.

This kind of workflow is definitely applicable to drawing.

Well, in 3d, even if I had started with individual bones, it would not be hard to adjust them afterwards. 3d is pretty forgiving when it comes to fiddling.

But this way is more succinct. I go from top to bottom, without needing to backtrack endlessly.

11am. Now focus me, I need to go further.

11:15am. Mhhh, let me do the sideways perspective.

11:30am. This is a pain in the ass. At any rate, the foot is the most difficult and I am done with it for now. Let me do the big bones. That should not be hard.

But my focus is low right now. Let me just do this and then I will leave it for the rest of the day until 6pm when I will do the chores. If I go to bed right after I finish the chores, I should get proper sleep. It is difficult for me to stop a late gaming session. So it makes sense to move it earlier.

12pm. I am yawing here. I've decided. I am quiting after dealing with the skeleton.

Art is too difficult after all, or more like - the reward of it is not enough to entice me. If I could tell a NN to just make a skeleton it would be an entirely different matter. If I had unlimited time, I could just grit my teeth and grind through it, but I have to prioritize.

I am just about returning to sanity after throwing in the towel last year. I am absolutely sure that I could not sustain this pace for art like I could for Spiral.

The only thing I am wondering about is music.

While I was doing the chores, I had the realiziation that I could do something synthetic, and have the NN do style transfer so it looks like it came from natural instruments. Something like that could be doable. I have no idea if it exists, I'd guess it does, but if it doesn't I could make it happen myself. If I had the money to train the NN to do it that is.

12:05pm. I am going to check music out. I want to satisfy my curiousity about its state. No way am I going to put in the level of effort I did into 3d though.

Let me get breakfast here. It is time for Bastard."