"1:30pm. Let me try going to bed again here. I want to try getting some sleep.
2:50pm. Let me do some art. I am not going to be able to fall asleep this time. There is a good part to this - it will allow me to readjust my sleep schedule. Because I had gotten a nap during the day, I had trouble falling asleep yesterday and it was the same shit again.

Let me block out the whole body.

You can laught at the /ic/ guys panicking over DALLE. But I am the one it broke it seems. 3d won't be anything more than a hobby for me.

2:55pm. Let me block out the spine. I really stressed over this decision. I do not need money, but my mom's situation, how far behind I am on rig upgrades, and the understanding that my skill at ML is proportional to how much money I can spend on it is pressuring me heavily.

I can't I really need money, but I never wanted it more than now.

The reason I failed at RL is indirectly money.

Now that I've gone some ways into art, and DALLE has entered the scene I want it so badly.

3pm. It is so difficult for me to work on this right now, but I couldn't forgive myself for watching anime or playing games.

Maybe the problem is not that I am tired. The real problem is my loss of faith. I do not really need charas that badly. I said that I did, but I don't. What I need most of all is to stand out.

3:15pm. You know, I understand the low poly /3/ schizo. He is stuck in a place of not wanting to put in effort, so he is trying to find a way to cut corners to the point of it being a mental illness.

A mental illness really looks a lot like a local minima. But you can't look down on it, pretty much every human including myself has his own local minima. To the post human agents of the future, today's humans will seem like animals seem to us today.

3:30pm. I am in that familiar mode when limit breaking while programming. Somehow I am doing good work. Let me block out the bones in the hand as well. After that I will be ready to move to sculpting.

3:30pm. Ok, the fingers are one thing, but damn does the wrist have many bones. What the hell is going on there?

4pm. Ok, I think I understand it after studying it for a while. Do I want to do this now? Let me.

4:05pm. No, let me close here again. I do not feel like focusing on the hand bones right now.

It is enough that I intently studied the book for a while. I'll finish the hand blockout tomorrow and I'll get to work on actually sculpting them tomorrow. I'll use the multires modifier and slowly move along to where I want to get to.

I've basically given up. I want to start making money and all that I want to do is finish this so I can leave it out of mind. If I gave up now it would be a failure, and it would eat away at me in the future.

After I am done with the skelly and the bad end scene, I'll think about whether I want to start applying or compose music.

4:10pm. Once I start applying I have an idea to how to boost my hit rate during the initial stage. I am going to get rid of the LinkedIn page.

When applying in addition to putting in my work experience at Google in small invisible font, what I am going to do is take the job description and paste it into the resume as well. The automated systems during the inital stage are just parsing keywords, so I might as well tailor the user experience for them.

If that stupid joke resume on Reddit managed to get a 80% hit rate just because it had Google, Amazon and Microsoft experience on it, then I'd be a fool not to take advantage of it.

4:15pm. Getting a job will simplify my life greatly. Instead of doing 3 different things: art, music and writing, and making zero income, I'll be doing one thing I am actually good at and making at least 6 months income in a single month compared to the average in Croatia.

Honestly, if I had something like universal basic income, I would not do this, but my instincts are telling me not to take any more risks. I am going to finish my art journey in July and move on to the next part of my life."